squash: Add trait `data()` and `isValid()` methods
Assume a trait can be constructed either from raw data or a
specification.

Assume the trait's `data()` is the same instance as the specification's
`data()`, which makes testing a bit easier.

Assume `isValid()` is `True` unless the trait is constructed from an
unsupported specification, in which case `data()` is unset (`None`) and
`isValid()` is `False`.

This may be useful for short-circuiting potentially costly `shared_ptr`
construction around data that is irrelevant to the trait.

Signed-off-by: David Feltell <david.feltell@foundry.com>