squash: Simplify, esp. combine SpecificationBase/Data
Combine `SpecificationBase` and `SpecificationData` into a single
`Specification` type using pimpl idiom to hide underlying
implementation. This centralises the data that needs transporting in
order to reproduce a specification, both it's properties and supported
traits.

`Specification`'s constructor takes a list of traits that it supports.
Initialise the underlying dict of a specification at construction with
an entry for each supported trait, as opposed to storing the trait list
independently. This is simpler but may be less performant - we should
profile before adding optimisations prematurely.

`BlobSpecification` is then simply a `Specification` with a defaulted
list of supported traits.

Bind `Specification` class to Python, allowing specifications to be
defined dynamically.

Add a `hasTrait` method to specifications which simply queries the
top-level keys of the underlying dict. Use this in the `isValid` method
of traits.

Signed-off-by: David Feltell <david.feltell@foundry.com>