[Tests] Split out "bridge" tests from "internal"
Part of #739. We were combining the "bridge" tests of the
`openassetio-python-bridge` with the "module" tests of Python bindings
helpers in `openassetio-python-module` and, somewhat arbitrarily,
grouping them under an "internal" heading.

These tests all relied on running under `pytest` and using an
`_openassetio_test` Python extension module to provide C++ helpers. This
makes sense for the "module" tests (`PyRetainingSharedPtr`) but was a
quick-and-dirty solution for the "bridge" tests
(`createPythonPluginSystemManagerImplementationFactory`).

So move the "bridge" tests out into their own directory with a C++
CTest that embeds a Python interpreter. This is much more representative
of how we expect the `openassetio-python-bridge` library to be used.

The leak sanitizer (via ASan) flags some leaks from pybind11 when the
test program exits. This is apparently a known bug
(https://github.com/pybind/pybind11/issues/3865). The leak arises when
defining a `py::enum_` and then calling `.value` on it to add an entry
to the enum.

It's mysterious that we've not seen it in previous sanitizer runs that
exercise the offending part of the code (i.e. `pytest` tests that import
the `_openassetio` bindings). It could have been hidden behind the
`python` executable, whereas for this test we have a custom
`openassetio-python-bridge-test-exe` executable that embeds Python. Or
we could be missing some cleanup in this executable, which `python`
handles properly.

Assuming this is a genuine bug, at least it's only a leak at program
exit, so is benign, and there's nothing we can do about it until
pybind11 fixes it upstream.

So just suppress any leaks from being reported if they have
`pybind11::enum_<` in their stack trace.

Signed-off-by: David Feltell <david.feltell@foundry.com>