[CI] Run sanitizers under RelWithDebInfo
Part of #739. In preparation for adding a new executable with known
third-party memory leaks, we need a way to suppress errors. LSan
supports suppression files that list patterns to search for in the
stack trace. So the stack trace must be available, meaning a Release
build won't cut it.

It's also useful to be able to see the stack trace if/when a leak is
detected, so its a good idea to run our sanitizer checks using a
RelWithDebInfo build regardless.

Signed-off-by: David Feltell <david.feltell@foundry.com>