[Tests] Assert getWithRelationship[s] copies default resultTraitSet
Closes #987. A common pitfall in pure Python is setting a default value
for an argument in a function to a mutable value, e.g.

```python
def someFunction(*args, resultTraitSet=set()):
  ...
```

In the above, any mutation to the default-provided `resultTraitSet` will
persist in the next invocation.

In our pybind11 bindings we have a similar pattern, i.e.

```c++
py::arg("resultTraitSet") = trait::TraitSet{}
```

where `TraitSet` is an alias for `std::unordered_set`.

It turns out that this is not a problem, since we're relying on
pybind11's automatic casting between a Python `set` and `unordered_set`,
which necessitates a copy of the default value into a new C++ value on
every invocation.

Stepping through in the debugger reveals that in pybind11 default
arguments are first converted to a Python object before being squirreled
away in a `function_record`, to be recalled when the function is called.

For `unordered_set`, this means a Python `set` is stored on the
`function_record`.

When calling the C++ function from Python, the squirreled away default
Python value is `cast`ed from the Python object back to a C++ object.

For `unordered_set` this is via a `set_caster`, which constructs a
`value` of type `unordered_set` (note: not a pointer or reference).

On the other hand, when a generic user defined type is used as a default
argument value, the standard `type_caster_generic` constructs a `value`
of type `void**`, which is filled in with a pointer to the value stored
in the `function_record` (which is itself heap-allocated and
move-constructed from the value given in the pybind11 `py::arg(...) =
...` call). Hence in the generic case, no copies are made.

So add a test to assert everything's OK. We may not need to keep this,
but it adds some confidence that if we change the `resultTraitSet`
argument's type, then we'll catch unexpected reference semantics.

Signed-off-by: David Feltell <david.feltell@foundry.com>