Hacky prototype of UI delegation
Adds a `ManagerUIBase` abstract base class with Python bindings.
Instances of this are returned by a `uiDelegate` function on the
`Manger`/`ManagerInterface`.  The class exposes a `getWidget` and
`setEntityChosenCallback`. The `getWidget` returns a `void*`. `std::any`
apparently doesn't work with pybind11 but wasn't pursued in detail.

Making the `ManagerUIBase` base class be a mixin to a widget class is
not possible in Python bindings (at least with PySide as the other base
class) - we get errors about conflicting metaclasses. Hence why a
`getWidget` method was needed, to separate the `ManagerUIBase`-derived
class from the widget.

Signed-off-by: David Feltell <david.feltell@foundry.com>