[Build] Do not override the libstdc++ ABI by default
Closes #1353. All VFX Reference Platform versions after CY22 specify
that the non-deprecated ABI for libstdc++ should be used. This means
that the default when building OpenAssetIO is wrong in those cases.

So in order to make a best guess at what the user wants when building
OpenAssetIO, simply don't enforce any ABI by default, so that the system
default is used.

The option remains available and works the same way for those that need
to be explicit.

This means CI diverges from CY22 unless we explicitly set the libstdc++
ABI. However, a CI is due to be upgraded to CY23+ in #812. So don't
bother forcing ABI only to change it again in the near future.

Signed-off-by: David Feltell <david.feltell@foundry.com>