[CI] Bump runners to approximate VFX CY24
Closes #812. Closes #1354. Part of #1389.

Bump/pin to Ubuntu 22.04 and Windows 2022 runners, which most closely
resemble VFX Reference Platform CY24. Similarly, bump the default Python
version used on CI to 3.11 to better match CY24.

For MacOS, there isn't a specific VFX Reference Platform requirement, so
leave as-is for now. That is, use the oldest version available on the
assumption that forward compatibility is more likely than backward
compatibility. MacOS will be futher addressed in #718.

Remove the boostrap script for macos-10.15 - it is unused.

Signed-off-by: David Feltell <david.feltell@foundry.com>