[Build] Don't export pybind11 exception types
Closes #1416.

Ideally, all OpenAssetIO dependencies are private and do not leak to the
host application (with the notable exception of Python itself). This
includes pybind11, where it would be ideal if OpenAssetIO's Python
bindings are independent of the host application, i.e. where the host
could use a different version of pybind11 if they wished, and everything
would still work.

However, it turns out that pybind11 exports it's exception class
symbols. In particular, the `pybind11::error_already_set` exception type
that is used to carry Python exception data through a C++ exception (and
potentially back to Python again).

For example, the ABI of the `error_already_set` type was changed
radically between pybind11 2.9 and 2.10. If OpenAssetIO is built with
2.10 and the host is built with 2.9, and the exception is caught in the
host (even as a `std::exception`), and `.what()` is called on it, the
wrong implementation of `.what()` may be executed, giving garbled text.
This exact scenario was seen in practice.

Disabling the export of the pybind11 exception will cause problems if we
ever need to catch a pybind11-specific exception in a different DSO. We
don't currently have that use-case, however.

So disable the export of pybind11 exceptions by overriding its
`PYBIND11_EXPORT_EXCEPTION` macro wherever the `pybind11::module` CMake
target is linked.

Host applications can still catch the pybind11 exception as a
`std::exception`, and safely call `.what()` on it to extract the error
message, and the correct implementation will be used.

Signed-off-by: David Feltell <david.feltell@foundry.com>