squash: Add skeleton C++ UI delegation API
Part of #1302. Add a new `openassetio-ui` target and C++ library. This
mostly duplicates code manager plugin code. That is, there is a
`UIDelegate` plugin system, manager-side interface, and host-side
middleware.

The interface and middleware a bare-bones without any useful methods.

The plugin system is more fleshed out. However, simply duplicating the
existing plugin system logic is problematic. The plugin system only
supports a single plugin per identifier. However, the UI delegation
design requires two plugins with the same identifier, one for the
manager plugin and one for the UI delegate plugin.

This is solved in the simplest way, by having a separate environment
variable for UI delegate search paths.

A better solution is to filter by plugin type when scanning. This
requires future work to redesign the `CppPluginSystem` (and
`PythonPluginSystem`) class. We assume for now that the user has
correctly set up their environment so plugin of the correct type are
located under the correct search path.

Signed-off-by: David Feltell <david.feltell@foundry.com>