squash: Add skeleton Python support for UI delegate plugins
Add the `openassetio-ui` library as a dependency of the `_openassetio`
Python extension module, and add bindings to it.

Ideally, UI delegation would be separated to its own Python package with
its own C extension module. However, disentangling the `bridge` library,
tests, and private utility code, is tricky. So in the spirit of
"batteries included" Python, assume for now that a single Python package
includes everything.

Add only basic tests for now, just some coverage of the UI delegate
plugin system.

Signed-off-by: David Feltell <david.feltell@foundry.com>