squash: Add UI state/request copy-constructing `make`
When stateChangedCallback is triggered, hosts may well want to be able
to compare the previous state with the new state. However, reference
semantics could mean that the new state _is_ the (mutated) old state. So
add an explicit way to deep copy `UIDelegateRequestPtr` and
`UIDelegateStatePtr`, using the same pattern as for `TraitsDataPtr`,
i.e. adding a `make()` overload, and binding it as a copy constructor in
Python.

Signed-off-by: David Feltell <david.feltell@foundry.com>