squash: Partially solve nativeData refcounting
One solution for `nativeData` refcounting is to abuse Python attributes
and `pybind11::caapsule` to add a deleter to Python wrappers, to ensure
the `PyObject` inside the `std::any` is properly decref'd when the
`UIDelegateState`/`Request` is destroyed.

This is a bit of a kludge, since we can't use the C++ destructor without
making `UIDelegateState`/`Request` polymorphic (in order to add a
trampoline class with a Python-specific destructor).

It also means we need to litter function signatures with
`PyRetainingSharedPtr` overloads to ensure the Python-only deleter
function is not lost (a process that is not complete as of this commit).

It also causes a wrinkle with C++ hosts that directly store a CPython
`PyObject*` in the `std::any` - cleanup of this becomes
non-deterministic.

It seems more straightforward to simply document that `nativeData` does
not own a reference, as dangerous as that is.

But keeping this solution in the commit history for posterity, before
refactoring.

Signed-off-by: David Feltell <david.feltell@foundry.com>