squash: WIP Attempt to use polymorphic request/state
Refactor the UIDelegateRequest/State to be stateful polymorphic
classes, with getters and setters. The idea is the Python bindings can
then add extra code to work around the issues of nativeData cleanup
and callback closure garbage collection (when there are circular
references).

However, this adds a lot more complexity and seemingly hacky fixes, and
won't currently compile due to some issue with the override macros and
returning reference to locals (probably needs a tweak to the
OPENASSETIO override macros).

Signed-off-by: David Feltell <david.feltell@foundry.com>