[Tests] Clarify generic test plugins vs. manager plugins
Part of #1445. Many tests do not actually need a full manager plugin,
and only require a generic `PythonPluginSystemPlugin` to be able to test
the required functionality.

In preparation for adding new plugin types (i.e. UI delegates), reduce
most test plugins to generic `PythonPluginSystemPlugin` subclasses,
and rename any that do need to be manager plugins to have "Manager"
somewhere in the name.

Signed-off-by: David Feltell <david.feltell@foundry.com>