[Core] Add CppPluginSystem validator callback
Part of #1445. We had filtering and error logic to further check
discovered `CppPluginSystemPlugin`s after the `scan()` process (i.e. to
ensure they were the appropriate subclass,
`CppPluginSystemManagerPlugin`). This was clunky and inefficient, and
not extensible.

So in preparation for adding more types of plugin (i.e. UI delegates),
and since we're already committed to breaking the API of
`CppPluginSystem.scan()`, add a validator callback that can be used
to filter out inappropriate plugins early, and can be customised by
the caller for different kinds of plugin.

Log such validation failures at "warning" level, in the expectation that
the module hook function name indicates the specific kind of plugin
that should be loaded, and so an invalid plugin class detected after
that point indicates plugin programmer error.

Signed-off-by: David Feltell <david.feltell@foundry.com>