[Python] Simplify Python binding for CppPluginSystemPlugin
Part of #1445.

The `CppPluginSystem` is exposed to Python, which was never really
necessary, since it's really an implementation detail of
`CppPluginSystemManagerImplementationFactory`. Having Python bindings
does help with tests (since we can test it through Python). But it means
we need to also bind `CppPluginSystemPlugin` (since instances of it
are returned by `CppPluginSystem`).

The `CppPluginSystemPlugin` class is only used (inherited from) by C++
plugins, and Python will never interact directly with these. It makes no
sense to construct (or inherit from) a `CppPluginSystemPlugin` in
Python.

So keep the binding of `CppPluginSystemPlugin`, but simplify to remove
the ability to construct or inherit from it. This simplifies tests and
clarifies its usage pattern.

Don't bother adding to RELEASE_NOTES since this was never supported
functionality.

Signed-off-by: David Feltell <david.feltell@foundry.com>