squash: Add CppPluginSystemUIDelegateImplementationFactory
Add a `CppPluginSystemUIDelegateImplementationFactory` that mirrors the
implementation of `CppPluginSystemManagerImplementationFactory`. This
class houses the logic required to load C++ UI delegate plugins.
Managers must implement their plugin class as a subclass of
`CppPluginSystemUIDelegatePlugin`, again mirroring the manager plugin
system. Ultimately, it is not expected that hosts will use this class
directly, other than constructing it, and will instead make use of
a `UIDelegateFactory`.

Signed-off-by: David Feltell <david.feltell@foundry.com>