squash: Add UIDelegateRequestInterface and UIDelegateStateInterface
Add `UIDelegateRequestInterface` for encapsulating the part of a request
for UI delegation that may change over time. For example, this could
include the selection of entities that the host wishes to present
additional information for via the UI delegate. Such a list could change
as the user interacts with the other parts of the host whilst the
delegated UI element is visible.

Similarly, add a `UIDelegateStateInterface` for encapsulating the
current state of a UI delegate. For example, the currently selected
entities in a delegated browser.

The request and state objects are also the mechanism for establishing
communication channels with an ongoing UI delegation request. This is
via the `stateChangedCallback` on the request, which is called by the UI
delegate when its (host-relevant) state has changed; and the
`updateRequestCallback` on the state, which is called by the host when
it wishes to update the parameters of the original request.

To remain agnostic to any particular UI framework, both the request and
state hold an arbitrary `nativeData` object. This is typically where a
widget object would be stored. Other possibilities include hosts that
use an immediate mode GUI framework, such as Imgui, using `nativeData`
to store a function that is called every frame. It is up to the host
application to document what is/should be placed in `nativeData` and
under what conditions. Documentation in UI-specific OpenAssetIO trait
libraries can provide further guidance on this.

Signed-off-by: David Feltell <david.feltell@foundry.com>