squash: Add "isPython" to UIDelegateInterface.info()
In the cross-language (C++/Python) case, where host xor UI delegate is
written in Python, the C++ side must ensure/expect `PyObject*` as the
`nativeData()` in `UIDelegateRequest` and `UIDelegateState`.

A C++ UI delegate can determine if the host is written in Python by
inspecting the `HostInterface.info()` (via `hostSession.host().info()`)
and checking if the `kInfoKey_IsPython` key is set.

Conversely, a C++ host must be able to determine if a UI delegate is
written in Python. So similarly add `kInfoKey_IsPython` to the default
`UIDelegateInterface.info()` dict in Python.

Signed-off-by: David Feltell <david.feltell@foundry.com>